Split into multiple files + type hints + update pre-commit
