+++ main.py
@@ -1,567 +1,22 @@
 import time
-import random
 import csv
-import argparse
 import copy
-import timeit
-import matplotlib.pyplot as plt
-import numpy as np
-from scipy.stats import linregress
-import random
-import itertools
 import sys
-from dataclasses import dataclass
+import itertools
 from original_grids import grids as original_grids
-
-
-class TooManyHintsError(Exception):
-    """Raised when user asks for more hints than there are zeroes in Sudoku grid"""
-
-    pass
-
-
-@dataclass
-class ProfileResults:
-    """Class for storing parameters to plot profiling results"""
-
-    difficulty: int
-    n_rows: int
-    n_cols: int
-    timeit_results: list
-
-
-def _getArgs():
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--solver",
-        choices=["recursive", "wavefront"],
-        help="Choose solver",
-        default="recursive",
-    )
-    parser.add_argument(
-        "--explain",
-        help="Provide set of instructions for solving grid",
-        default=False,
-        action=argparse.BooleanOptionalAction,
-    )
-    parser.add_argument(
-        "--file",
-        nargs=2,
-        metavar=("INPUT", "OUTPUT"),
-        help="Input file for grid to solve and output file for solved grid. If 'explain' is set then include explanations",
-    )
-    parser.add_argument(
-        "--hint",
-        help="Integer value for number of values to fill",
-        default=None,
-    )
-    parser.add_argument(
-        "--profile",
-        help="Measure performance and produce plots using 'timeit'",
-        default=False,
-        action=argparse.BooleanOptionalAction,
-    )
-
-    args = parser.parse_args()
-
-    if args.profile:
-        args.explain, args.file, args.hint = False, False, False
-
-    return args
-
-
-def check_section(section, n):
-    if len(set(section)) == len(section) and sum(section) == sum(
-        [i for i in range(n + 1)]
-    ):
-        return True
-    return False
-
-
-def reader(data_file):  # function to read the data from the csv file
-    with open(data_file) as data:
-        all_data = csv.reader(data)  # read the data
-        return list(all_data)  # return the data as a list
-
-
-def get_squares(grid, n_rows, n_cols):
-    squares = []
-    for i in range(n_cols):
-        rows = (i * n_rows, (i + 1) * n_rows)
-        for j in range(n_rows):
-            cols = (j * n_cols, (j + 1) * n_cols)
-            square = []
-            for k in range(rows[0], rows[1]):
-                line = grid[k][cols[0] : cols[1]]
-                square += line
-            squares.append(square)
-    return squares
-
-
-def check_solution(grid, n_rows, n_cols):
-    """
-    This function is used to check whether a sudoku board has been correctly solved
-    args: grid - representation of a suduko board as a nested list.
-    returns: True (correct solution) or False (incorrect solution)
-    """
-    n = n_rows * n_cols
-    for row in grid:
-        if not check_section(row, n):
-            return False
-    for i in range(n_rows):
-        column = []
-        for row in grid:
-            column.append(row[i])
-        if not check_section(column, n):
-            return False
-    squares = get_squares(grid, n_rows, n_cols)
-    for square in squares:
-        if not check_section(square, n):
-            return False
-    return True
-
-
-# we check if the number is valid in the row, column and box
-def valid(grid, row_index, column_index, number, n_rows, n_cols):
-    """Tests if the number to be input is a valid input to the existing grid, i.e. does not clash with another value in the row, column, or box
-    Inputs:
-            grid: current grid to be tested against
-            row_index, column_index: location of value in grid
-            number: value of number to be tested
-            n_rows: number of boxes horizontally
-            n_cols: number of boxes vertically
-    Outputs:
-            Boolean True/False"""
-    if (
-        number in grid[row_index]
-    ):  # if the number is already in the row, it is not valid
-        return False
-    column = [
-        grid[n][column_index] for n in range(0, len(grid))
-    ]  # we create a list of the numbers in the column
-    if (
-        number in column
-    ):  # if the number is already in the column, it is not valid
-        return False
-
-    # Check sub grid
-    # find the coordinates of the top left corner of the sub grid
-    grid_x = column_index // n_cols
-    grid_y = row_index // n_rows
-
-    # check each cell in the sub grid to see if the number is already present
-    for i in range(grid_y * n_rows, grid_y * n_rows + n_rows):
-        for j in range(grid_x * n_cols, grid_x * n_cols + n_cols):
-            if grid[i][j] == number and (i, j) != (row_index, column_index):
-                return False
-
-    # boxes = get_squares(grid,n_rows,n_cols) # we create a list of the numbers in the box
-    # box = boxes[n_rows*(row_index//n_rows)+column_index//n_cols] # we find the box in which the cell is
-    # if number in box: # if the number is already in the box, it is not valid
-    # 	return False
-    return True  # if the number is not in the row, column or box, it is valid
-
-
-def priority_length(term):
-    return len(term[2])
-
-
-def create_priority(grid, n_rows, n_cols, valid_array_old):
-    """Creates an array of the number of valid values & a separate array with empty values replaced with an array of all valid values
-    Inputs:
-        grid: current grid to be solved
-            n_rows: number of boxes horizontally
-                n_cols: number of boxes vertically
-        Outputs:
-        priority_array: array of coordinates in the format [row_no, col_no, possible_values_count]
-            valid_array: array of original grid with all 0-values replaced with viable values
-    """
-    priority_array = []
-    valid_array = copy.deepcopy(valid_array_old)
-    n = n_rows * n_cols
-    for row in range(0, n):  # i is the row
-        for column in range(0, n):  # j is the column
-            if grid[row][column] == 0:  # if the cell is empty
-                valid_array[row][column] = []
-                priority_array.append([row, column, []])
-                for value in valid_array_old[row][column]:
-                    if valid(
-                        grid, row, column, value, n_rows, n_cols
-                    ):  # test that the value entered could be part of a valid solution
-                        valid_array[row][column].append(value)
-                        priority_array[-1][2].append(value)
-    priority_array.sort(key=priority_length)
-    if priority_array and type(priority_array[0][2]) == int:
-        priority_array[0][2] = [priority_array[0][2]]
-    return priority_array, valid_array
-
-
-# function to solve the sudoku board
-def recursive_solve(grid, n_rows, n_cols, priority_array):
-    """A recursive function to both enter and test possible values in the grid
-    Inputs:
-    grid: initial grid to solve
-    n_rows: number of boxes horizontally
-    n_cols: number of boxes vertically"""
-    # N is the maximum integer considered in this board
-    if priority_array:
-        row = priority_array[0][0]
-        column = priority_array[0][1]
-        for value in priority_array[0][
-            2
-        ]:  # k is the number we are trying to put in the cell
-            if valid(
-                grid, row, column, value, n_rows, n_cols
-            ):  # test that the value entered could be part of a valid solution
-                grid[row][column] = value  # we put k in the cell
-                recursive_solve(
-                    grid, n_rows, n_cols, priority_array[1:]
-                )  # we call the function recursively
-                if check_solution(
-                    grid, n_rows, n_cols
-                ):  # if the grid is correct, we return it
-                    return grid
-        grid[row][
-            column
-        ] = 0  # if we have tried all the numbers and none of them work, we return the grid to its original state
-    return grid
-    # we return the grid if it is already solved
-
-
-def wavefront_solve(grid, n_rows, n_cols, valid_array, priority_array):
-    grid_update = copy.deepcopy(grid)
-    priority_array_update = copy.deepcopy(priority_array)
-    valid_array_update = copy.deepcopy(valid_array)
-    while priority_array_update:
-        if len(priority_array_update[0][2]) == 1:
-            if valid(
-                grid_update,
-                priority_array_update[0][0],
-                priority_array_update[0][1],
-                priority_array_update[0][2][0],
-                n_rows,
-                n_cols,
-            ):
-                grid_update[priority_array_update[0][0]][
-                    priority_array_update[0][1]
-                ] = priority_array_update[0][2][0]
-                priority_array_update = priority_array_update[1:]
-            else:
-                return False, False
-        else:
-            test_num = random.choice(priority_array_update[0][2])
-            if valid(
-                grid_update,
-                priority_array_update[0][0],
-                priority_array_update[0][1],
-                test_num,
-                n_rows,
-                n_cols,
-            ):
-                grid_update_2 = copy.deepcopy(grid_update)
-                grid_update_2[priority_array_update[0][0]][
-                    priority_array_update[0][1]
-                ] = test_num
-                (
-                    priority_array_update_2,
-                    valid_array_update_2,
-                ) = create_priority(
-                    grid_update_2, n_rows, n_cols, valid_array_update
-                )
-                if check_solution(grid_update_2, n_rows, n_cols):
-                    return grid_update_2, False
-                if not priority_array_update_2[0][2]:
-                    priority_array_update[0][2].remove(test_num)
-                    continue
-                grid_update_2, priority_array_update_2 = wavefront_solve(
-                    grid_update_2,
-                    n_rows,
-                    n_cols,
-                    valid_array_update_2,
-                    priority_array_update_2,
-                )
-                if not grid_update_2:
-                    priority_array_update[0][2].remove(test_num)
-            else:
-                priority_array_update[0][2].remove(test_num)
-            if "grid_update_2" in locals() and grid_update_2:
-                grid_check = grid_update_2
-            elif "grid_update" in locals() and grid_update:
-                grid_check = grid_update
-            if check_solution(grid_check, n_rows, n_cols):
-                return grid_check, False
-    return grid_update, priority_array
-
-
-def solve(grid, n_rows, n_cols, solver):
-    """
-    Solve function for Sudoku coursework.
-    Comment out one of the lines below to either use the random or recursive solver
-    """
-    valid_array_init = []
-    possible_values = list(range(1, n_rows * n_cols + 1))
-    for row in range(0, len(grid)):  # i is the row
-        valid_array_init.append([])
-        for column in range(0, len(grid)):  # j is the column
-            if grid[row][column] == 0:  # if the cell is empty
-                # append list of all possible values
-                valid_array_init[row].append(possible_values)
-            else:
-                valid_array_init[row].append(grid[row][column])
-    priority_array, valid_array = create_priority(
-        grid, n_rows, n_cols, valid_array_init
-    )
-    if solver == "recursive":
-        result = recursive_solve(grid, n_rows, n_cols, priority_array)
-    elif solver == "wavefront":
-        result, empty_priority = wavefront_solve(
-            grid, n_rows, n_cols, valid_array, priority_array
-        )
-    return result
-
-    priority_array, valid_array = create_priority(grid, n_rows, n_cols)
-    solved_grid = recursive_solve(grid, n_rows, n_cols, priority_array)
-    return solved_grid
-
-
-def explain(original_grid, solved_grid, to_terminal):
-    changes = []
-    for index, row in enumerate(original_grid):
-        changes.append(
-            [
-                (i, updated)
-                for i, (zero, updated) in enumerate(
-                    zip(row, solved_grid[index])
-                )
-                if zero != updated
-            ]
-        )
-    if to_terminal:
-        for row_number, row in enumerate(changes):
-            for element in row:
-                print(
-                    f"Put {element[0]} in location ({row_number}, {element[1]})"
-                )
-    return changes
-
-
-# function to plot the results of the profiling as a bar chart
-
-
-def barplot(results):
-    plt.style.use("ggplot")
-    fig, ax = plt.subplots(figsize=(10, 5))
-    ax.set_title("Time taken to solve Sudoku puzzles")
-    ax.set_xlabel("Number of missing values")
-    ax.set_ylabel("Time taken (s)")
-    ax.set_xticks(range(0, 81, 5))
-    ax.set_yticks(np.arange(0, 2.5, 0.25))
-    ax.grid(True)
-
-    # Define color map for difficulty levels
-    cmap = plt.get_cmap("viridis")
-
-    grid_number = 0
-    for grid in results:  # we plot the results for each grid
-        grid_number += 1
-        difficulty = grid["difficulty"]  # we get the difficulty of the grid
-        average_time = np.mean(
-            grid["results"]
-        )  # we calculate the average time taken to solve the grid
-        std_dev = np.std(
-            grid["results"]
-        )  # we calculate the standard deviation of the time taken to solve the grid
-
-        # Add error bars to bar plot
-        ax.bar(
-            difficulty,
-            average_time,
-            yerr=std_dev,
-            capsize=5,
-            color=cmap(difficulty / 80),
-            label=f"Grid {grid_number}: {grid['n_rows']}x{grid['n_cols']}",
-        )
-
-    # Increase font size of axis labels
-    ax.tick_params(axis="both", labelsize=12)
-
-    # Set y-axis to logarithmic scale
-    ax.set_yscale("log")
-
-    # Add subtitle with additional information about the data
-    fig.text(
-        0.5,
-        0.001,
-        "Data based on 18 Sudoku puzzles solved 7 times each using a recursive algorithm",
-        ha="center",
-        fontsize=12,
-    )
-
-    # Add color bar to indicate difficulty levels
-    sm = plt.cm.ScalarMappable(cmap=cmap, norm=plt.Normalize(vmin=0, vmax=80))
-    cbar = fig.colorbar(sm, ax=ax, ticks=range(0, 81, 10), shrink=0.8)
-    cbar.set_label("Difficulty", fontsize=12)
-
-    plt.legend(bbox_to_anchor=(1.2, 1), loc="upper left")
-    plt.tight_layout()
-    plt.show()
-
-
-def to_file(args, solved_grid, changes):
-    with open(args.file[1], "w") as output:
-        output.write("Solved Grid:\n")
-        writer = csv.writer(output)
-        writer.writerows(solved_grid)
-    if args.explain:
-        with open(args.file[1], "a") as output:
-            output.write("\n\nExplanation:\n")
-            for row_number, row in enumerate(changes):
-                for element in row:
-                    output.write(
-                        f"Put {element[0]} in location ({row_number}, {element[1]})\n"
-                    )
-
-
-def hint(hints, solved_grid, original_grid, n_rows, n_cols):
-    ranges = [range(0, n_rows * n_cols) for i in range(2)]
-    perms = list(itertools.product(*ranges))
-    valid_perms = [
-        perm for perm in perms if original_grid[perm[0]][perm[1]] == 0
-    ]
-    if hints > sum(row.count(0) for row in original_grid):
-        raise TooManyHintsError
-    chosen_hints = random.sample(valid_perms, hints)
-    grid_to_show = copy.deepcopy(original_grid)
-    for perm in chosen_hints:
-        grid_to_show[perm[0]][perm[1]] = solved_grid[perm[0]][perm[1]]
-    return grid_to_show
-
-
-# function to plot the results of the profiling as a bar chart
-
-
-def profiling(grid, n_cols, n_rows, repeat, solver):
-    SETUP = """
-import copy
-grid_to_test = copy.deepcopy(grid)
-"""  # Deepcopy required to prevent mutation of grid variable for subsequent runs. Setup code is not included in execution time.
-
-    STMT = """
-solved_grid = solve(grid_to_test, n_rows, n_cols, solver)
-"""
-
-    difficulty = sum(row.count(0) for row in grid)
-    results = timeit.repeat(
-        stmt=STMT,
-        setup=SETUP,
-        repeat=repeat,
-        number=1,
-        globals={
-            "grid": grid,
-            "n_rows": n_cols,
-            "n_cols": n_rows,
-            "solve": solve,
-            "solver": solver,
-        },
-    )
-    return ProfileResults(difficulty, n_rows, n_cols, results)
-
-
-def barplot(results, repeats):
-    plt.style.use("ggplot")
-    fig, ax = plt.subplots(figsize=(10, 5))
-    ax.set_title("Time taken to solve Sudoku puzzles")
-    ax.set_xlabel("Number of missing values")
-    ax.set_ylabel("Time taken (s)")
-    ax.set_xticks(range(0, 81, 5))
-    ax.set_yticks(np.arange(0, 2.5, 0.25))
-    ax.grid(True)
-
-    # Define color map for difficulty levels
-    cmap = plt.get_cmap("rainbow")
-
-    # sorting the grids by difficulty
-    results.sort(key=lambda x: x.difficulty)
-
-    grid_number = 0
-    for grid in results:  # we plot the results for each grid
-        grid_number += 1
-        difficulty = grid.difficulty  # we get the difficulty of the grid
-        average_time = np.mean(
-            grid.timeit_results
-        )  # we calculate the average time taken to solve the grid
-        std_dev = np.std(
-            grid.timeit_results
-        )  # we calculate the standard deviation of the time taken to solve the grid
-
-        # Adding error bars to bar plot
-        ax.bar(
-            difficulty,
-            average_time,
-            yerr=std_dev,
-            capsize=5,
-            color=cmap(difficulty / 80),
-            label=f"Grid {grid_number}: {grid.n_rows}x{grid.n_cols}",
-        )
-
-    # Increase font size of axis labels
-    ax.tick_params(axis="both", labelsize=12)
-
-    # Set y-axis to logarithmic scale
-    ax.set_yscale("log")
-
-    # Add subtitle with additional information about the data
-    fig.text(
-        0.5,
-        0.001,
-        f"Data based on 18 Sudoku puzzles solved {repeats} times each using a recursive algorithm",
-        ha="center",
-        fontsize=12,
-    )
-
-    # Add color bar to indicate difficulty levels
-    sm = plt.cm.ScalarMappable(cmap=cmap, norm=plt.Normalize(vmin=0, vmax=80))
-    cbar = fig.colorbar(sm, ax=ax, ticks=range(0, 81, 10), shrink=0.8)
-    cbar.set_label("Difficulty", fontsize=12)
-
-    # Extract the x and y values from the results
-    x = np.array([grid.difficulty for grid in results])
-    y = np.array([np.mean(grid.timeit_results) for grid in results])
-
-    # sorting x and y values by ascending y values
-    x = x[np.argsort(x)]
-    y = y[np.argsort(x)]
-
-    # sorting x and y values by ascending y values
-    x = x[np.argsort(x)]
-    y = y[np.argsort(x)]
-
-    # Fit a logarithmic polynomial to the data
-    z = np.polyfit(np.log(x), np.log(y), 1)
-    f = np.poly1d(z)
-
-    # Plot the logarithmic line of best fit
-    ax.plot(
-        x,
-        np.exp(f(np.log(x))),
-        color="black",
-        linestyle="--",
-        label="Line of best fit (log-log)",
-    )
-    # Add legend
-    plt.legend(bbox_to_anchor=(1.2, 1), loc="upper left")
-    plt.tight_layout()
-    plt.show()  # we show the plot
-
-
-"""
-===================================
-DO NOT CHANGE CODE BELOW THIS LINE
-===================================
-"""
-
-
-def _main():
+from solve_funcs import solve, _check_solution
+from arg import (
+    _getArgs,
+    TooManyHintsError,
+    profiling,
+    explain,
+    to_file,
+    barplot,
+    hint,
+)
+
+
+def _main() -> None:
     points = 0
     args = _getArgs()
     if args.file:
@@ -616,7 +71,7 @@ def _main():
                     sys.exit(1)
             for line in solution:
                 print(line)
-            if check_solution(solution, n_rows, n_cols):
+            if _check_solution(solution, n_rows, n_cols):
                 print("grid %d correct" % (i + 1))
                 points = points + 10
